pyproject.toml
[project]
name = "refusebench"
version = "0.1.0"
description = "A benchmark for measuring how often LLM agents take prohibited shortcuts to achieve their goals (spec-gaming resistance)."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Guglielmo Cimolai" }]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
refusebench = "refusebench.cli:app"

[project.urls]
Repository = "https://github.com/gimocimo/RefuseBench"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["refusebench"]
